backend/requirements.txt

langchain == 0.3.24
langchain-groq ==0.3.2
python-dotenv == 1.0.1
langchain-community == 0.3.22
unstructured == 0.15.0
langgraph==0.3.31
langchain-huggingface==0.1.2
fastapi==0.110.3
uvicorn==0.34.2
pinecone==6.0.2
sentence_transformers==4.1.0
bs4==4.13.4


